chore: remove deprecated task redux code
