--- src/js/tasks/api.js
@@ -1,5 +1,0 @@
-import { Request } from "../app/request";
-
-export const list = () => Request.get("/tasks");
-
-export const get = ({ taskId }) => Request.get(`/tasks/${taskId}`);
